Change the direction of pings to allow for ENR adding
A first ping will allow the passing of ENR of client b through the
discv5 handshake and adding it to the routing table of client a.

Making the node seen should typically not be required ot use the
node to send a FindContent request, but perhaps different
implementations don't allow that.